requirements.txt
streamlit
google-generativeai
python-dotenv

PyPDF2
pypdf
pdf2image

chromadb
faiss-cpu

groq
langchain
langchain-groq
langchain_google_genai
langchain_community

youtube_transcript_api
pathlib